Finished application
Added music and sound to the application and the ability to toggle those. Also added better documentation. In the end, we couldn't cleanly replicate the card flipping animation, even though we were close so we removed the all code related to it. Then removed font-file as there were some issues loading it. Overall a good project to practice effects, refs, and http requests again.